+++ environment/environment_templates.go
@@ -5,7 +5,7 @@ import (
 	"path/filepath"
 
 	"github.com/smartcontractkit/integrations-framework/config"
-	"github.com/smartcontractkit/integrations-framework/internal"
+	"github.com/smartcontractkit/integrations-framework/tools"
 	coreV1 "k8s.io/api/core/v1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
@@ -21,8 +21,8 @@ const (
 func NewAdapterManifest() *K8sManifest {
 	return &K8sManifest{
 		id:             "adapter",
-		DeploymentFile: filepath.Join(internal.ProjectRoot, "/environment/templates/adapter-deployment.yml"),
-		ServiceFile:    filepath.Join(internal.ProjectRoot, "/environment/templates/adapter-service.yml"),
+		DeploymentFile: filepath.Join(tools.ProjectRoot, "/environment/templates/adapter-deployment.yml"),
+		ServiceFile:    filepath.Join(tools.ProjectRoot, "/environment/templates/adapter-service.yml"),
 
 		values: map[string]interface{}{
 			"apiPort": AdapterAPIPort,
@@ -50,8 +50,8 @@ func NewAdapterManifest() *K8sManifest {
 func NewChainlinkManifest() *K8sManifest {
 	return &K8sManifest{
 		id:             "chainlink",
-		DeploymentFile: filepath.Join(internal.ProjectRoot, "/environment/templates/chainlink-deployment.yml"),
-		ServiceFile:    filepath.Join(internal.ProjectRoot, "/environment/templates/chainlink-service.yml"),
+		DeploymentFile: filepath.Join(tools.ProjectRoot, "/environment/templates/chainlink-deployment.yml"),
+		ServiceFile:    filepath.Join(tools.ProjectRoot, "/environment/templates/chainlink-service.yml"),
 
 		values: map[string]interface{}{
 			"webPort": ChainlinkWebPort,
@@ -75,8 +75,8 @@ func NewChainlinkManifest() *K8sManifest {
 func NewHardhatManifest() *K8sManifest {
 	return &K8sManifest{
 		id:             "hardhat",
-		DeploymentFile: filepath.Join(internal.ProjectRoot, "/environment/templates/hardhat-deployment.yml"),
-		ServiceFile:    filepath.Join(internal.ProjectRoot, "/environment/templates/hardhat-service.yml"),
+		DeploymentFile: filepath.Join(tools.ProjectRoot, "/environment/templates/hardhat-deployment.yml"),
+		ServiceFile:    filepath.Join(tools.ProjectRoot, "/environment/templates/hardhat-service.yml"),
 
 		values: map[string]interface{}{
 			"rpcPort": EVMRPCPort,
